apps/python-flask/requirements.txt
Flask>=3.0.0,<4.0.0
gunicorn>=22.0.0,<24.0.0
azure-monitor-opentelemetry>=1.6.0,<2.0.0
opentelemetry-api>=1.25.0,<2.0.0
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
